// zwangine-designer/src/core/model/workflow-definition.ts
export class ErrorHandlerDefinition extends WorkflowElement {
    stepName?: string = 'errorHandler';
    id?: string = 'errorHandler-' + uuidv4().substring(0,4);
    deadLetterChannel?: DeadLetterChannelDefinition;
    defaultErrorHandler?: DefaultErrorHandlerDefinition;
    jtaTransactionErrorHandler?: JtaTransactionErrorHandlerDefinition;
    noErrorHandler?: NoErrorHandlerDefinition;
    refErrorHandler?: RefErrorHandlerDefinition | string;
    springTransactionErrorHandler?: SpringTransactionErrorHandlerDefinition;
    public constructor(init?: Partial<ErrorHandlerDefinition>) {
        super('ErrorHandlerDefinition');
        Object.assign(this, init);
    }
}

export class ExpressionSubElementDefinition extends WorkflowElement {
    stepName?: string = 'expressionSubElement';
    constant?: ConstantExpression | string;
    csimple?: CSimpleExpression | string;
    datasonnet?: DatasonnetExpression | string;
    exchangeProperty?: ExchangePropertyExpression | string;
    groovy?: GroovyExpression | string;
    header?: HeaderExpression | string;
    hl7terser?: Hl7TerserExpression | string;
    java?: JavaExpression | string;
    jq?: JqExpression | string;
    js?: JavaScriptExpression | string;
    jsonpath?: JsonPathExpression | string;
    language?: LanguageExpression;
    method?: MethodCallExpression | string;
    mvel?: MvelExpression | string;
    ognl?: OgnlExpression | string;
    python?: PythonExpression | string;
    ref?: RefExpression | string;
    simple?: SimpleExpression | string;
    spel?: SpELExpression | string;
    tokenize?: TokenizerExpression | string;
    variable?: VariableExpression | string;
    wasm?: WasmExpression | string;
    xpath?: XPathExpression | string;
    xquery?: XQueryExpression | string;
    xtokenize?: XMLTokenizerExpression | string;
    public constructor(init?: Partial<ExpressionSubElementDefinition>) {
        super('ExpressionSubElementDefinition');
        Object.assign(this, init);
    }
}
// ...
export class FromDefinition extends WorkflowElement {
    stepName?: string = 'from';
    id?: string = 'from-' + uuidv4().substring(0,4);
    description?: string;
    uri: string = '';
    variableReceive?: string;
    parameters?: any = {};
    steps: WorkflowElement[] = [];
    public constructor(init?: Partial<FromDefinition>) {
        super('FromDefinition');
        Object.assign(this, init);
    }
}
// ...
export class InputTypeDefinition extends WorkflowElement {
    stepName?: string = 'inputType';
    id?: string = 'inputType-' + uuidv4().substring(0,4);
    description?: string;
    urn: string = '';
    validate?: boolean;
    public constructor(init?: Partial<InputTypeDefinition>) {
        super('InputTypeDefinition');
        Object.assign(this, init);
    }
}

export class InterceptDefinition extends WorkflowElement {
    stepName?: string = 'intercept';
    id?: string = 'intercept-' + uuidv4().substring(0,4);
    description?: string;
    disabled?: boolean;
    onWhen?: OnWhenDefinition;
    steps?: WorkflowElement[] = [];
    public constructor(init?: Partial<InterceptDefinition>) {
        super('InterceptDefinition');
        Object.assign(this, init);
    }
}

export class InterceptFromDefinition extends WorkflowElement {
    stepName?: string = 'interceptFrom';
    id?: string = 'interceptFrom-' + uuidv4().substring(0,4);
    description?: string;
    disabled?: boolean;
    uri?: string;
    onWhen?: OnWhenDefinition;
    steps?: WorkflowElement[] = [];
    public constructor(init?: Partial<InterceptFromDefinition>) {
        super('InterceptFromDefinition');
        Object.assign(this, init);
    }
}

export class InterceptSendToEndpointDefinition extends WorkflowElement {
    stepName?: string = 'interceptSendToEndpoint';
    id?: string = 'interceptSendToEndpoint-' + uuidv4().substring(0,4);
    description?: string;
    disabled?: boolean;
    uri: string = '';
    skipSendToOriginalEndpoint?: string;
    afterUri?: string;
    onWhen?: OnWhenDefinition;
    steps?: WorkflowElement[] = [];
    public constructor(init?: Partial<InterceptSendToEndpointDefinition>) {
        super('InterceptSendToEndpointDefinition');
        Object.assign(this, init);
    }
}
// ...
export class OnCompletionDefinition extends WorkflowElement {
    stepName?: string = 'onCompletion';
    id?: string = 'onCompletion-' + uuidv4().substring(0,4);
    description?: string;
    disabled?: boolean;
    mode?: string;
    onCompleteOnly?: boolean;
    onFailureOnly?: boolean;
    parallelProcessing?: boolean;
    executorService?: string;
    useOriginalMessage?: boolean;
    onWhen?: OnWhenDefinition;
    steps?: WorkflowElement[] = [];
    public constructor(init?: Partial<OnCompletionDefinition>) {
        super('OnCompletionDefinition');
        Object.assign(this, init);
    }
}

export class OnExceptionDefinition extends WorkflowElement {
    stepName?: string = 'onException';
    id?: string = 'onException-' + uuidv4().substring(0,4);
    description?: string;
    disabled?: boolean;
    exception?: string[] = [];
    onWhen?: OnWhenDefinition;
    retryWhile?: ExpressionSubElementDefinition;
    redeliveryPolicy?: RedeliveryPolicyDefinition;
    redeliveryPolicyRef?: string;
    handled?: ExpressionSubElementDefinition;
    continued?: ExpressionSubElementDefinition;
    onRedeliveryRef?: string;
    onExceptionOccurredRef?: string;
    useOriginalMessage?: boolean;
    useOriginalBody?: boolean;
    steps?: WorkflowElement[] = [];
    public constructor(init?: Partial<OnExceptionDefinition>) {
        super('OnExceptionDefinition');
        Object.assign(this, init);
    }
}
// ...
export class OnWhenDefinition extends WorkflowElement {
    stepName?: string = 'onWhen';
    id?: string = 'onWhen-' + uuidv4().substring(0,4);
    description?: string;
    expression?: ExpressionDefinition;
    public constructor(init?: Partial<OnWhenDefinition>) {
        super('OnWhenDefinition');
        Object.assign(this, init);
    }
}
// ...
export class OutputTypeDefinition extends WorkflowElement {
    stepName?: string = 'outputType';
    id?: string = 'outputType-' + uuidv4().substring(0,4);
    description?: string;
    urn: string = '';
    validate?: boolean;
    public constructor(init?: Partial<OutputTypeDefinition>) {
        super('OutputTypeDefinition');
        Object.assign(this, init);
    }
}
// ...
export class RouteConfigurationDefinition extends WorkflowElement {
    stepName?: string = 'routeConfiguration';
    id?: string;
    description?: string;
    errorHandler?: ErrorHandlerDefinition;
    intercept?: InterceptDefinition[] = [];
    interceptFrom?: InterceptFromDefinition[] = [];
    interceptSendToEndpoint?: InterceptSendToEndpointDefinition[] = [];
    onException?: OnExceptionDefinition[] = [];
    onCompletion?: OnCompletionDefinition[] = [];
    precondition?: string;
    public constructor(init?: Partial<RouteConfigurationDefinition>) {
        super('RouteConfigurationDefinition');
        Object.assign(this, init);
    }
}

export class RouteContextRefDefinition extends WorkflowElement {
    stepName?: string = 'routeContextRef';
    ref: string = '';
    public constructor(init?: Partial<RouteContextRefDefinition>) {
        super('RouteContextRefDefinition');
        Object.assign(this, init);
    }
}

export class RouteDefinition extends WorkflowElement {
    stepName?: string = 'route';
    id?: string = 'route-' + uuidv4().substring(0,4);
    description?: string;
    group?: string;
    nodePrefixId?: string;
    routeConfigurationId?: string;
    autoStartup?: boolean;
    startupOrder?: number;
    streamCache?: boolean;
    trace?: boolean;
    messageHistory?: boolean;
    logMask?: boolean;
    errorHandlerRef?: string;
    shutdownRoute?: string;
    shutdownRunningTask?: string;
    precondition?: string;
    errorHandler?: ErrorHandlerDefinition;
    inputType?: InputTypeDefinition;
    outputType?: OutputTypeDefinition;
    from: FromDefinition = new FromDefinition();
    public constructor(init?: Partial<RouteDefinition>) {
        super('RouteDefinition');
        Object.assign(this, init);
    }
}
// ...
export class DeadLetterChannelDefinition extends WorkflowElement {
    stepName?: string = 'deadLetterChannel';
    id?: string = 'deadLetterChannel-' + uuidv4().substring(0,4);
    deadLetterUri: string = '';
    deadLetterHandleNewException?: boolean;
    redeliveryPolicy?: RedeliveryPolicyDefinition;
    useOriginalMessage?: boolean;
    useOriginalBody?: boolean;
    redeliveryPolicyRef?: string;
    loggerRef?: string;
    level?: string;
    logName?: string;
    onRedeliveryRef?: string;
    onExceptionOccurredRef?: string;
    onPrepareFailureRef?: string;
    retryWhileRef?: string;
    executorServiceRef?: string;
    public constructor(init?: Partial<DeadLetterChannelDefinition>) {
        super('DeadLetterChannelDefinition');
        Object.assign(this, init);
    }
}

export class DefaultErrorHandlerDefinition extends WorkflowElement {
    stepName?: string = 'defaultErrorHandler';
    id?: string = 'defaultErrorHandler-' + uuidv4().substring(0,4);
    redeliveryPolicy?: RedeliveryPolicyDefinition;
    useOriginalMessage?: boolean;
    useOriginalBody?: boolean;
    redeliveryPolicyRef?: string;
    loggerRef?: string;
    level?: string;
    logName?: string;
    onRedeliveryRef?: string;
    onExceptionOccurredRef?: string;
    onPrepareFailureRef?: string;
    retryWhileRef?: string;
    executorServiceRef?: string;
    public constructor(init?: Partial<DefaultErrorHandlerDefinition>) {
        super('DefaultErrorHandlerDefinition');
        Object.assign(this, init);
    }
}

export class JtaTransactionErrorHandlerDefinition extends WorkflowElement {
    stepName?: string = 'jtaTransactionErrorHandler';
    id?: string = 'jtaTransactionErrorHandler-' + uuidv4().substring(0,4);
    transactedPolicyRef?: string;
    rollbackLoggingLevel?: string;
    redeliveryPolicy?: RedeliveryPolicyDefinition;
    useOriginalMessage?: boolean;
    useOriginalBody?: boolean;
    redeliveryPolicyRef?: string;
    loggerRef?: string;
    level?: string;
    logName?: string;
    onRedeliveryRef?: string;
    onExceptionOccurredRef?: string;
    onPrepareFailureRef?: string;
    retryWhileRef?: string;
    executorServiceRef?: string;
    public constructor(init?: Partial<JtaTransactionErrorHandlerDefinition>) {
        super('JtaTransactionErrorHandlerDefinition');
        Object.assign(this, init);
    }
}

export class NoErrorHandlerDefinition extends WorkflowElement {
    stepName?: string = 'noErrorHandler';
    id?: string = 'noErrorHandler-' + uuidv4().substring(0,4);
    public constructor(init?: Partial<NoErrorHandlerDefinition>) {
        super('NoErrorHandlerDefinition');
        Object.assign(this, init);
    }
}

export class RefErrorHandlerDefinition extends WorkflowElement {
    stepName?: string = 'refErrorHandler';
    id?: string = 'refErrorHandler-' + uuidv4().substring(0,4);
    ref: string = '';
    public constructor(init?: Partial<RefErrorHandlerDefinition>) {
        super('RefErrorHandlerDefinition');
        Object.assign(this, init);
    }
}

export class SpringTransactionErrorHandlerDefinition extends WorkflowElement {
    stepName?: string = 'springTransactionErrorHandler';
    id?: string = 'springTransactionErrorHandler-' + uuidv4().substring(0,4);
    transactedPolicyRef?: string;
    rollbackLoggingLevel?: string;
    redeliveryPolicy?: RedeliveryPolicyDefinition;
    useOriginalMessage?: boolean;
    useOriginalBody?: boolean;
    redeliveryPolicyRef?: string;
    loggerRef?: string;
    level?: string;
    logName?: string;
    onRedeliveryRef?: string;
    onExceptionOccurredRef?: string;
    onPrepareFailureRef?: string;
    retryWhileRef?: string;
    executorServiceRef?: string;
    public constructor(init?: Partial<SpringTransactionErrorHandlerDefinition>) {
        super('SpringTransactionErrorHandlerDefinition');
        Object.assign(this, init);
    }
}